Started to move towards a cleaner parser.
The parser is responsible for all parsing, regardless of whether its the
main template, our data from output markup or tags. The main benefit is less
duplication. Want to parse a value for either an output tag or an assign tag?
use parse.ReadValue().

Some other benefits include:
- More context aware (we can give the line number and a good view of where
  errors happened)
- Less objects being created (we create far less slices, and just keep a
  position of where we are in the data)
- Easier to test

The downside is that it'll probably be a bit slower (way more method calls),
but that might be offset by the the fewer slices being created.