Basic comparison logic. Still missing contains though.
Need to play with this to see if it works out. Not sure how much of the type
system to hide. It's not that I'm against hiding it, it's that there are
some clear cases where it leaks, and having some hidden with these edge cases
can make it pretty inconsistent.

For example:

"124" == 124   // true
"124.0" == 124 // false

It's solvable, but is it right to solve it. Moreso, it's already and quickly
feeling pretty ugly in there.